docs: 게시글 검색 관련 API 문서 수정
